Move Observation to its own file
