+++ profiling/src/profile/internal/mod.rs
@@ -1,0 +1,5 @@
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
+
+mod observation;
+pub use observation::*;
